various tweaks to jordyd's updates.

+++ Source/NSIndexPath+Etoile.m
@@ -79,7 +79,7 @@ Will raise an NSInvalidArgumentException if the separator is nil. */
     NSString *path = [NSString stringWithFormat: @"%lu", (unsigned long)[self firstIndex]];
     NSUInteger indexCount = [self length];
 
-    for (int i = 1; i < indexCount; i++)
+    for (NSUInteger i = 1; i < indexCount; i++)
     {
         path = [path stringByAppendingString:
             [NSString stringWithFormat: @"%@%lu", separator, (unsigned long)[self indexAtPosition: i]]];
